# Use struct instead of tuple to define UnpackedPath